Replace QScrollArea with QTextEdit
As we're not using show the generated image anymore, I decided to use a
QTextEdit for storing log messages of the algorithm's steps, which
actually looks like a better idea.

There are also two new buttons, one opens the image with the feh
command utility and the other saves the log messages to any given file
name.

Signed-off-by: Paulo Alcantara <pcacjr@gmail.com>